[refactor] Move garbage collection related code into a separate file
